integration test updated cross checked, but must fail on GitHub automation

+++ install-asciidoc.sh
@@ -1,20 +1,24 @@
-#!/usr/bin/env bash
+#!/bin/bash
 #
-# This shell script was automatically compiled from install-asciidoc.sh.jam
-# It is used to automate the installation of the development version of jamal-asciidoc plugin
+# This script is used to automate the installation of the development version of jamal-asciidoc plugin
 # from inside the development environment.
 #
 echo "Compiling and installing Jamal, the whole project"
-mvn clean install
+if ! mvn clean install ;then
+  exit 1
+fi
+
+VERSION=$(mvn help:evaluate -Dexpression=project.version -q -DforceStdout)
+
 echo "Creating the JAMAL_PROJECT_HOME/.asciidoctor/lib directory"
 mkdir -p .asciidoctor/lib
 echo "cd into JAMAL_PROJECT_HOME/.asciidoctor/lib directory"
-pushd .asciidoctor/lib
+cd .asciidoctor/lib || exit 1
 echo "removing all old files"
 rm -f *
 echo "unzipping the libraries from the local maven repo"
-unzip ~/.m2/repository/com/javax0/jamal/jamal-asciidoc/2.8.3-SNAPSHOT/jamal-asciidoc-2.8.3-SNAPSHOT-jamal-asciidoc-distribution.zip >/dev/null
+unzip ~/.m2/repository/com/javax0/jamal/jamal-asciidoc/$VERSION/jamal-asciidoc-$VERSION-jamal-asciidoc-distribution.zip >/dev/null
 echo "cd back to Jamal main project directory"
-popd
+cd ../.. || exit 1
 echo "Restart IntelliJ, then you have the new version"
 echo "DONE"
